+++ server/src/main/java/com/goose/crosstimer/crossroad/service/CrossroadService.java
@@ -1,5 +1,6 @@
 package com.goose.crosstimer.crossroad.service;
 
+import com.goose.crosstimer.api.dto.TDataCrossroadResponse;
 import com.goose.crosstimer.api.dto.TDataSignalResponse;
 import com.goose.crosstimer.api.service.TDataApiService;
 import com.goose.crosstimer.common.exception.CustomException;
@@ -7,6 +8,7 @@ import com.goose.crosstimer.common.exception.ErrorCode;
 import com.goose.crosstimer.crossroad.domain.Crossroad;
 import com.goose.crosstimer.crossroad.dto.CrossroadRangeRequest;
 import com.goose.crosstimer.crossroad.dto.CrossroadRangeResponse;
+import com.goose.crosstimer.crossroad.mapper.CrossroadMapper;
 import com.goose.crosstimer.signal.domain.SignalCache;
 import com.goose.crosstimer.crossroad.dto.CrossroadWithSignalResponse;
 import com.goose.crosstimer.crossroad.repository.CrossroadJpaRepository;
@@ -22,7 +24,9 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
@@ -37,10 +41,12 @@ public class CrossroadService {
     private final SignalCacheMapper signalCacheMapper;
     private final RedissonClient redissonClient;
 
-    private static final long LOCK_WAIT_TIME_MS = 10; //Lock 대기 시작
-    private static final long LOCK_LEASE_TIME_MS = 2_000; //Lock 최대 점유 시간
+    private static final long LOCK_WAIT_MS = 10; //Lock 대기 시작
+    private static final long LOCK_LEASE_MS = 2_000; //Lock 최대 점유 시간
     private static final long MAX_CACHE_WAIT_MS = 150; //Cache polling 최대 주기
     private static final long POLL_INTERVAL_MS = 20; //Cache polling 주기
+
+    private static final String UPSERT_LOCK_KEY = "lock:crossroad:monthlyUpsert";
 
     public CrossroadWithSignalResponse getCrossroadWithSignals(Integer crossroadId) {
         log.debug("getCrossroadWithSignals 호출: crossroadId={}", crossroadId);
@@ -78,7 +84,7 @@ public class CrossroadService {
         boolean locked = false;
 
         try {
-            locked = rlock.tryLock(LOCK_WAIT_TIME_MS, LOCK_LEASE_TIME_MS, TimeUnit.MILLISECONDS);
+            locked = rlock.tryLock(LOCK_WAIT_MS, LOCK_LEASE_MS, TimeUnit.MILLISECONDS);
             if (locked) {
                 //double check
                 optionalCache = signalCacheRepository.findById(crossroadId);
@@ -106,7 +112,6 @@ public class CrossroadService {
             }
         }
     }
-
 
     private SignalCache fetchAndCacheSignal(Integer crossroadId) {
         //외부 API 호출
@@ -179,5 +184,55 @@ public class CrossroadService {
                 crossroad.getLat(),
                 crossroad.getLng()
         );
+    }
+
+    /**
+     * TData 교차로 MAP 정보 API를 호출하여 제공된 모든 교차로 정보를 MySQL에 Upsert 처리
+     */
+    @Transactional
+    public void upsertCrossroads() {
+        RLock lock = redissonClient.getLock(UPSERT_LOCK_KEY);
+        boolean lockAcquired = false;
+        try {
+            lockAcquired = lock.tryLock(0, TimeUnit.MINUTES);
+            if (lockAcquired) {
+                log.info("교차로 데이터 Upsert");
+
+                List<TDataCrossroadResponse> fetched = new ArrayList<>();
+                int pageNo = 1;
+                while (true) {
+                    List<TDataCrossroadResponse> responses = tDataApiService.getCrossroadsMaxRow(pageNo++);
+                    if (responses.isEmpty()) { //더 이상 조회 안될때까지 반복
+                        break;
+                    }
+                    fetched.addAll(responses);
+                }
+
+                int saveCount = 0;
+                int updateCount = 0;
+                for (TDataCrossroadResponse response : fetched) {
+                    Optional<Crossroad> findCrossroad = crossroadJpaRepository.findById(response.crossroadId());
+
+                    if (findCrossroad.isEmpty()) { //신규 교차로인 경우 insert
+                        crossroadJpaRepository.save(CrossroadMapper.fromDto(response));
+                        saveCount++;
+                    } else { // 기존 교차로인 경우 update
+                        findCrossroad.get().update(
+                                response.name(),
+                                response.lat(),
+                                response.lng()
+                        );
+                        updateCount++;
+                    }
+                }
+                log.info("교차로 데이터 Upsert 완료 : save {}, update {}", saveCount, updateCount);
+            }
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        } finally {
+            if (lockAcquired) {
+                lock.unlock();
+            }
+        }
     }
 }
